修复 新闻页面状态栏颜色的问题（临时） 优化 页面状态 新增 教室查询页面日期选择器 去除 设置中无用的选项 优化 设置软件外链接 新增 数据储存

+++ app/src/main/java/com/xhand/hnu/model/entity/JwEntity.kt
@@ -3,7 +3,7 @@ package com.xhand.hnu.model.entity
 // 返回参数
 data class LoginEntity(
     val msg: String,
-    val code: Long,
+    val code: Int,
     val user: UserEntity? = null,
 )
